requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
google-generativeai==0.8.3
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.0
